+++ sdk/core/azure_core/src/fs/mod.rs
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+//! Asynchronous file system utilities.
+
+#[cfg(feature = "tokio")]
+pub use typespec_client_core::fs::*;