+++ sdk/core/azure_core/src/http/options/mod.rs
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+mod telemetry;
+
+pub use telemetry::*;
+pub use typespec_client_core::http::{
+    ClientMethodOptions, ClientOptions, ExponentialRetryOptions, FixedRetryOptions, RetryOptions,
+    TransportOptions,
+};