+++ sdk/core/azure_core/src/http/request/mod.rs
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+//! HTTP requests.
+
+pub mod options;
+
+pub use typespec_client_core::http::request::*;