+++ sdk/identity/azure_identity/src/credentials/virtual_machine_managed_identity_credential.rs
@@ -4,8 +4,7 @@
 use crate::{ImdsId, ImdsManagedIdentityCredential, TokenCredentialOptions};
 use azure_core::{
     credentials::{AccessToken, TokenCredential},
-    headers::HeaderName,
-    Url,
+    http::{headers::HeaderName, Url},
 };
 use std::sync::Arc;
 
@@ -43,9 +42,5 @@ impl VirtualMachineManagedIdentityCredential {
 impl TokenCredential for VirtualMachineManagedIdentityCredential {
     async fn get_token(&self, scopes: &[&str]) -> azure_core::Result<AccessToken> {
         self.credential.get_token(scopes).await
-    }
-
-    async fn clear_cache(&self) -> azure_core::Result<()> {
-        self.credential.clear_cache().await
     }
 }