+++ sdk/typespec/typespec_client_core/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Release History
 
+## 0.2.0 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+- Consolidated all the `tokio` features into a single feature named `tokio`. Traits remain separate but `tokio` support is enabled with a single feature.
+- Removed `Header` re-export from `http` module. It is still defined in the `http::headers` module.
+- Removed `Pager`.
+- Removed `parsing` module.
+
+### Bugs Fixed
+
+### Other Changes
+
+- Use `std::sync::LazyLock` added in rustc 1.80 instead of `once_cell::sync::Lazy`.
+
 ## 0.1.0 (2025-02-18)
 
 ### Features Added